#6: Did some changes.
