+++ src/main/java/org/weathertrack/Main.java
@@ -18,5 +18,6 @@ public class Main {
 	);
 
 	public static void main(String[] args) {
+
 	}
 }
