+++ src/main/java/org/weathertrack/api/service/forecast/ForecastApiService.java
@@ -1,4 +1,5 @@
 package org.weathertrack.api.service.forecast;
 
 public interface ForecastApiService {
+	// TODO: Implement methods, should also return ResponseData
 }
