+++ src/main/java/org/weathertrack/forecast/ForecastProvider.java
@@ -1,4 +1,5 @@
 package org.weathertrack.forecast;
 
 public interface ForecastProvider {
+	// TODO: Implement methods, should also return ResponseData
 }
