+++ src/main/java/org/weathertrack/geocoding/GeocodingProvider.java
@@ -1,4 +1,5 @@
 package org.weathertrack.geocoding;
 
 public interface GeocodingProvider {
+	// TODO: Implement methods, should also return ResponseData
 }
